requirements.txt
requests==2.22.0
json-tools==0.4.1
PyMySQL==1.0.2
PyYAML==5.1.1
